backend/requirements.txt
asgiref==3.7.2
black==23.9.1
click==8.1.7
colorama==0.4.6
Django==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
Pillow==10.0.1
platformdirs==3.11.0
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
django-cors-headers==3.14.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-telegram-bot==20.6
asyncio==3.4.3
